Use an `off_t` type to hold the raw size of an ELF object.
Ticket:		#568

git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@3711 95820547-d848-0410-985e-9ae8fe0fa350

+++ libelf/_libelf.h
@@ -90,7 +90,7 @@ struct _Elf {
 	Elf_Kind	e_kind;		/* ELF_K_* */
 	Elf		*e_parent; 	/* non-NULL for archive members */
 	unsigned char	*e_rawfile;	/* uninterpreted bytes */
-	size_t		e_rawsize;	/* size of uninterpreted bytes */
+	off_t		e_rawsize;	/* size of uninterpreted bytes */
 	unsigned int	e_version;	/* file version */
 
 	/*
